@@ -1,8 +1,8 @@
 module metrics
 
-go 1.22.12
+go 1.23.0
 
-toolchain go1.22.12
+toolchain go1.24.1
 
 require (
 	github.com/go-chi/chi/v5 v5.2.1
